feat(TvDemo): implement windowWidth reactive property and resize event listener for responsive behavior
