+++ src/composable/useDemo.js
@@ -18,13 +18,8 @@ const useDemo = (props) => {
   const viewportHeight = ref(360);
   const viewportWidth = ref('100%');
   const scrollTop = ref(0);
-  const windowWidth = ref(typeof window !== 'undefined' ? window.innerWidth : 1200);
   let resizeObserver = null;
   let fallbackResizeListenerAttached = false;
-
-  const updateWindowWidth = () => {
-    windowWidth.value = window.innerWidth;
-  };
 
   const variantEntries = computed(() =>
     (props.variants || []).map((variant, index) => ({
@@ -148,7 +143,6 @@ const useDemo = (props) => {
       }
 
       isMounted.value = true;
-      window.addEventListener('resize', updateWindowWidth);
     }
 
     nextTick(updateViewportHeight);
@@ -160,9 +154,6 @@ const useDemo = (props) => {
       resizeObserver = null;
     }
     detachFallbackResizeListener();
-    if (typeof window !== 'undefined') {
-      window.removeEventListener('resize', updateWindowWidth);
-    }
   });
 
   const fetchReadme = async () => {
@@ -446,7 +437,6 @@ const useDemo = (props) => {
     addLog,
     clearLogs,
     viewportWidth,
-    windowWidth,
   };
 };
